+++ pkg/actions/tools/ffmpeg/demuxer.go
@@ -7,8 +7,12 @@ import (
 	"github.com/rsteube/carapace"
 )
 
+// ActionDemuxers completes demuxers
+//
+//	aax (CRI AAX)
+//	ac3 (raw AC-3)
 func ActionDemuxers() carapace.Action {
-	return carapace.ActionExecCommand("ffmpeg", "-demuxers")(func(output []byte) carapace.Action {
+	return carapace.ActionExecCommand("ffmpeg", "-hide_banner", "-demuxers")(func(output []byte) carapace.Action {
 		lines := strings.Split(string(output), "\n")
 		r := regexp.MustCompile(`^ (?P<type>.).{1} (?P<name>[^ ]+) +(?P<description>.*)$`)
