+++ pkg/actions/tools/ffmpeg/format.go
@@ -1,4 +1,4 @@
-package action
+package ffmpeg
 
 import (
 	"regexp"
@@ -7,10 +7,14 @@ import (
 	"github.com/rsteube/carapace"
 )
 
-func ActionDevices() carapace.Action {
-	return carapace.ActionExecCommand("ffmpeg", "-hide_banner", "-devices")(func(output []byte) carapace.Action {
+// ActionFormats completes formats
+//
+//	aax (CRI AAX)
+//	ac3 (raw AC-3)
+func ActionFormats() carapace.Action {
+	return carapace.ActionExecCommand("ffmpeg", "-hide_banner", "-formats")(func(output []byte) carapace.Action {
 		lines := strings.Split(string(output), "\n")
-		r := regexp.MustCompile(`^.{3} (?P<devices>[^ ]+) +(?P<description>.*)$`)
+		r := regexp.MustCompile(`^.{3} (?P<format>[^ ]+) +(?P<description>.*)$`)
 
 		vals := make([]string, 0)
 		for _, line := range lines[4 : len(lines)-1] {
